Custom Feature creating done. Testing with application context :thumbsup:

+++ src/main/java/org/scripps/branch/repository/UserRepository.java
@@ -7,6 +7,6 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
 	public User findByEmail(String email);
 	
-	public User findById(int id);
+	public User findById(long id);
 
 }
